[#665] Big renaming: class2_Camera, class3_Crystal, class4_grating, class5_Diagnostic
