+++ tofu/data/_class5_plot.py
@@ -10,6 +10,7 @@ import matplotlib.pyplot as plt
 import matplotlib.colors as mcolors
 import datastock as ds
 
+
 # specific
 from . import _mesh_comp
 from . import _generic_check
